load_service/requirements.txt
psycopg2-binary==2.9.9
prometheus-client==0.19.0 